==> 1 Collect_EODHD.ipynb <==
import pandas as pd
import requests
from datetime import date, datetime
from dateutil import rrule
from pandas.io.json import json_normalize

# Create list with companies we want to loop through
comp_list = ["AAPL",
             "MSFT",
             "GOOG",
             "AMZN",
             "PCAR",
             "TSLA",
             "NVDA",
             "META",
             "AVGO",
             "ASML",
             "PEP",
             "AZN",
             "COST",
             "CSCO",
             "TMUS", 
             "ADBE",
             "TXN",
             "CMCSA",
             "NFLX",
             "AMD"]

# Create list with years we want to loop through
# We can max 
start_date_list = []
end_date_list = []
 
# dates
start_date = datetime(2021, 1, 1)
end_date = datetime(2024, 1, 1)
 
for dt in rrule.rrule(rrule.YEARLY, dtstart=start_date, until=end_date):
    start_date_list.append(dt.strftime("%Y-%m-%d"))
    end_date_list.append(dt.strftime("%Y-%m-%d"))
    
start_date_list = list(start_date_list[:-1])
end_date_list = list(end_date_list[1:])
end_date_list[3]=datetime(2023, 3, 20).strftime("%Y-%m-%d")

news_df = pd.DataFrame()

for comp in comp_list:
    for start, end in zip(start_date_list, end_date_list):
    # Collect news using API
        r = requests.get("https://eodhistoricaldata.com/api/news?api_token="
                         + #APIKEY
                         + "&from=" # Note that API key is removed
                         + str(start) + "&to="
                         + str(end) + "&s="
                         + str(comp) 
                         + str("&offset=0&limit=1000"))
    
        responses = r.json()
    
        # Define collected API responses
        data = pd.DataFrame(responses)
        data = data.drop_duplicates('title', keep = 'first')
    
        data["company"] = comp
        news_df = news_df.append(data)
        
news_df

# Save as CSV
news_df.to_csv("EODHD.csv", index = False)
