==> 1 Collect_mediastack.ipynb <==
import pandas as pd
import requests
from datetime import datetime
from datetime import date
from dateutil import rrule
from pandas.io.json import json_normalize

# Create list with companies we want to loop through
comp_list = ["Apple AAPL",
             "Microsoft MSFT",
             "Alphabet GOOG",
             "Amazon AMZN",
             "PACCAR PCAR",
             "Tesla TSLA",
             "NVIDIA NVDA",
             "Facebook META",
             "Broadcom AVGO",
             "ASML",
             "PEP",
             "AZN",
             "COST",
             "CSCO",
             "TMUS", 
             "ADBE",
             "TXN",
             "CMCSA",
             "NFLX",
             "AMD"]

# Create list with months we want to loop through
# We can max 
start_date_list = []
end_date_list = []
 
# dates
start_date = datetime(2020, 1, 1)
end_date = datetime(2023, 3, 20)
 
for dt in rrule.rrule(rrule.MONTHLY, dtstart=start_date, until=end_date):
    start_date_list.append(dt.strftime("%Y-%m-%d"))
    end_date_list.append(dt.strftime("%Y-%m-%d"))
    
start_date_list = list(start_date_list[:-1])
end_date_list = list(end_date_list[1:])

news_df = pd.DataFrame()

for comp in comp_list:
    for start, end in zip(start_date_list, end_date_list):
        # Collect news using API
        r = requests.get("https://api.mediastack.com/v1/news?access_key="
                        + #API KEY
                        + "&date=" 
                        + str(start) + ","
                        + str(end)
                        + "&languages=en"
                        + "&countres=gb,us"
                        + "&categories=business,technology,science&keywords="
                        + str(comp) 
                        + "&limit=100")
        
        responses = r.json()
        
        #Define collected API responses
        data = responses['data']
        
        #Normalize and unnest JSON response
        flatData = json_normalize(data)
        
        df = pd.DataFrame(flatData)
        df["company"] = comp
        news_df = news_df.append(df)
        
news_df

# Save as CSV
#news_df.to_csv("mediastack.csv", index = False)
