==> 1 Collect_tweets.ipynb <==
# Import packages for scraping tweets
import pandas as pd
from datetime import date, timedelta
from tqdm.notebook import tqdm
import snscrape.modules.twitter as sntwitter

# Create range of dates to loop through
def daterange(start_date, end_date):
    for n in range(int((end_date - start_date).days)):
        yield start_date + timedelta(n)

start_date = date(2021, 1, 1)
end_date = date(2023, 3, 20)

## Create list with companies we want to loop through
comp_list = ["$AAPL",
             "$MSFT",
             "$GOOG",
             "$AMZN",
             "$PCAR",
             "$TSLA",
             "$NVDA",
             "$META",
             "$AVGO",
             "$ASML",
             "$PEP",
             "$AZN",
             "$COST",
             "$CSCO",
             "$TMUS",  
             "$ADBE",
             "$TXN",
             "$CMCSA",
             "$NFLX",
             "$AMD"]

# Creates list to save tweets
tweets = []

# Create loop to scrape daily 
for company in comp_list:
    for single_date in daterange(start_date, end_date):
        scraper = sntwitter.TwitterSearchScraper(str(company) +
                                                 " since:" + 
                                                 str(single_date) + 
                                                 " until:" +
                                                 str(single_date + timedelta(1)) +
                                                 " lang:en exclude:replies exclude:retweets")
        for i, tweet in enumerate(scraper.get_items()):
            if i < 3:
                tweets.append([tweet.date,
                               tweet.rawContent,
                               company])
                
            else:
                break
    
    

tweets_df = pd.DataFrame(tweets, columns=["date", "content", "company"])

# Save as CSV
tweets_df.to_csv("tweets.csv", index = False)
