==> 2 Merge & Format.ipynb <==
# Load Packages
import pandas as pd
from datetime import date

# Load Tweets
tweets = pd.read_csv("tweets.csv",
                     lineterminator='\n')

# Rearrange Data
tweets = tweets[["company", "date", "content"]] # Select Columns
tweets['date'] = pd.to_datetime(tweets['date']) # Format Date
tweets['date'] = tweets['date'].dt.date 
tweet

# Load News Articles from Mediastack & EODHD
mediastack = pd.read_csv("mediastack.csv",
                         lineterminator='\n')

EODHD = pd.read_csv("EODHD.csv",
                    lineterminator='\n')

EODHD2 = pd.read_csv("EODHD2.csv",
                    lineterminator='\n')

# Rearrange Data
mediastack = mediastack[["company", "published_at", "description"]] # Select Columns
mediastack.columns = ["company", "date", "content"] # Rename
mediastack['date'] = pd.to_datetime(mediastack['date']) # Format Date
mediastack['date'] = mediastack['date'].dt.date
mediastack["company"] = mediastack["company"].str[:4].str.upper() # Rename company such that equal to company ticker

EODHD = EODHD[["company", "date", "content"]] # Select columns
EODHD['date'] = pd.to_datetime(EODHD['date']) # Format Date
EODHD['date'] = EODHD['date'].dt.date

EODHD2 = EODHD2[["company", "date", "content"]] # Select columns
EODHD2['date'] = pd.to_datetime(EODHD2['date']) # Format Date
EODHD2['date'] = EODHD2['date'].dt.date

news = mediastack.append(EODHD)
news = news.append(EODHD2)
news

# Save as CSV
tweets.to_csv("tweets_formated.csv", index = False)
news.to_csv("news_formated.csv", index = False)
