==> 3 Sentiment.ipynb <==
# Import packages needed for text cleaning
import pandas as pd
import numpy as np
import re
import nltk
from nltk.stem import WordNetLemmatizer
  
lemmatizer = WordNetLemmatizer()
nltk.download('words')
words = set(nltk.corpus.words.words())

# Load data 
tweets = pd.read_csv("tweets_formated.csv",
            lineterminator='\n')

news = pd.read_csv("news_formated.csv",
            lineterminator='\n')

# If we check the two dataframes, we see that 'news' contains more companies
# Example: AMAZ = AMZN, APPL = AAPL
tweets.groupby(['company'])['company'].count() # 10 companies
news.groupby(['company'])['company'].count() # 18 companies -> using wrong tic

# Check if the news articles are indeed about Amazon
news.loc[news['company'] == 'AMAZ']
news.loc[news['company'] == 'ALPH']
news.loc[news['company'] == 'APPL']
news.loc[news['company'] == 'BROA']
news.loc[news['company'] == 'FACE']
news.loc[news['company'] == 'MICR']
news.loc[news['company'] == 'NVID']
news.loc[news['company'] == 'PACC']
news.loc[news['company'] == 'TESL']

# Rename to correct company names
news = news.replace("AMAZ", "AMZN")
news = news.loc[news["company"] != "ALPH"]
news = news.replace("APPL", "AAPL")
news = news.replace("BROA", "AVGO")
news = news.replace("FACE", "META")
news = news.replace("MICR", "MSFT")
news = news.replace("NVID", "NVDA")
news = news.replace("PACC", "PCAR")
news = news.replace("TESL", "TSLA")

tweets['company'] = tweets['company'].str[1:]

# Print ticker, company name, and number of tweets and news for Table
print(tweets.groupby(['company'])['company'].count())
print(news.groupby(['company'])['company'].count())

# Function to clean text, more specifically tweets
def clean_text(text):
    if type(text) == np.float:
        return ""
    temp = text.lower()
    temp = re.sub("'", "", temp) # To avoid removing contractions in english
    temp = re.sub("@[A-Za-z0-9_]+","", temp) # Removes @user
    temp = re.sub(r'http\S+', '', temp) # Removes http links
    temp = re.sub('[()!?]', ' ', temp) 
    temp = re.sub('\[.*?\]',' ', temp)
    temp = re.sub("[^a-z0-9]"," ", temp)
    temp = temp.split() 
    text = text.replace("#", "").replace("_", " ") #Remove hashtag sign but keep the text
    temp = " ".join(word for word in temp)
    return temp

tweets['content'] = tweets['content'].map(lambda x: clean_text(x))
news['content'] = news['content'].map(lambda x: clean_text(x))

# Tokenize data
def tokenization(text):
    text = re.split('\W+', text)
    return text

tweets['tokenized'] = tweets['content'].apply(lambda x: tokenization(x.lower()))
news['tokenized'] = news['content'].apply(lambda x: tokenization(x.lower()))

# Removes stopwords
#nltk.download()
from nltk.corpus import stopwords
stopword = nltk.corpus.stopwords.words('english')

def remove_stopwords(text):
    text = [word for word in text if word not in stopword]
    return text
    
tweets['nonstop'] = tweets['tokenized'].apply(lambda x: remove_stopwords(x))
news['nonstop'] = news['tokenized'].apply(lambda x: remove_stopwords(x))

# Stemming the text
ps = nltk.PorterStemmer()

def stemming(text):
    text = [ps.stem(word) for word in text]
    return text

tweets['stemmed'] = tweets['nonstop'].apply(lambda x: stemming(x))
news['stemmed'] = news['nonstop'].apply(lambda x: stemming(x))

# Loughram & McDonald sentiment analysis
import pysentiment2 as ps

lm = ps.LM()
tweets['polarity'] = tweets['stemmed'].apply(lm.get_score)
news['polarity'] = news['stemmed'].apply(lm.get_score)

# Change data structure
tweets = pd.concat(
    [tweets.drop(['polarity'], axis=1), 
     tweets['polarity'].apply(pd.Series)], axis=1)

news = pd.concat(
    [news.drop(['polarity'], axis=1), 
     news['polarity'].apply(pd.Series)], axis=1)

# Create new variable with sentiment "neutral," "positive" and "negative"
tweets['sentiment'] = tweets['Polarity'].apply(lambda x: 'positive' if x > 0 else 'neutral' if x == 0 else 'negative')
news['sentiment'] = news['Polarity'].apply(lambda x: 'positive' if x > 0 else 'neutral' if x == 0 else 'negative')

# Calculate sentiment mean for each day
tweets_polarity = tweets.groupby(["company", "date"])["Polarity"].mean()
news_ = news.groupby(["company", "date"])["Polarity"].mean()

# Change data stracture back 
tweets_sent = pd.DataFrame(tweets_polarity)
news_sent = pd.DataFrame(news_polarity)

tweets_sent = tweets_sent.reset_index()
news_sent = news_sent.reset_index()

# Save as CSV
tweets_sent.to_csv("tweets_sent.csv", index = False)
news_sent.to_csv("news_sent.csv", index = False)
